+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 
 ## Version 2
 
+### Version 2.2.2
+
+* [154](https://github.com/mlebreuil/netbox-contract/issues/154) Fix edit and delete bulk operations on dimensions and invoice lines.
+* [153](https://github.com/mlebreuil/netbox-contract/issues/153) Enforce uniquness of accounting dimensions.
+* Adds a status ( Active or Inactive ) to accounitng dimensions.
+* [151](https://github.com/mlebreuil/netbox-contract/issues/151) Fix accounting line and dimensions search.
+
 ### Version 2.2.1
 
 * [142](https://github.com/mlebreuil/netbox-contract/issues/142) Gives the option to enter contract yearly recuring costs instead of only monthly recuring costs.